+++ src/lib/components/forms/form/example/form-fields.ts
@@ -106,6 +106,13 @@ export const containerCompontent: TerraFormFieldInterface = {
 };
 
 export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
+    checkbox: {
+        type: 'checkbox',
+        options: {
+            name: 'checkbox',
+            required: true
+        }
+    },
     textarea: {
         type: 'textarea',
         options: {