+++ src/lib/components/forms/form/model/form-type-map.ts
@@ -1,7 +1,6 @@
 import { Type } from '@angular/core';
 import { CheckboxGroupComponent } from '../../checkbox-group/checkbox-group.component';
 import { TerraFormTypeInterface } from './terra-form-type.interface';
-import { TerraCheckboxComponent } from '../../checkbox/terra-checkbox.component';
 import { TerraDatePickerComponent } from '../../input/date-picker/terra-date-picker.component';
 import { TerraFileInputComponent } from '../../input/file-input/terra-file-input.component';
 import { TerraTextInputComponent } from '../../input/text-input/terra-text-input.component';
@@ -15,6 +14,7 @@ import { TerraCodeEditorComponent } from '../../../editors/code-editor/terra-cod
 import { TerraSliderComponent } from '../../slider/terra-slider.component';
 import { TerraPortletComponent } from '../../../layouts/portlet/terra-portlet.component';
 import { TerraSuggestionBoxComponent } from '../../suggestion-box/terra-suggestion-box.component';
+import { CheckboxComponent } from '../form-components/checkbox/checkbox.component';
 import { TextAreaComponent } from '../form-components/text-area/text-area.component';
 
 /**
@@ -24,9 +24,9 @@ import { TextAreaComponent } from '../form-components/text-area/text-area.compon
 export class FormTypeMap {
     public readonly checkbox: TerraFormTypeInterface = {
         inputMap: {
-            name: 'inputCaption'
+            required: 'isRequired'
         },
-        component: TerraCheckboxComponent
+        component: CheckboxComponent
     };
     public readonly date: TerraFormTypeInterface = {
         inputMap: {